FEATURE Make compiler output suppressable.
system() is now called with intern = TRUE. The compiler output
is saved to a local variable. By default, this variable is not
printed to command line. The output can be printed by setting
verbose = TRUE. This argument must be added to all functions
which allow compiling. It is already present in dMod::Y() and
dMod::generateModel().